+++ readFile.cpp
@@ -13,6 +13,7 @@ void questions::openFile(ifstream& inputStream){
     
 }
 
+// gets the whole next line
 void questions::setFullQuestion(ifstream& inputStream){
     questions::fullQuestion.clear();
     getline(inputStream, questions::fullQuestion);
@@ -20,9 +21,10 @@ void questions::setFullQuestion(ifstream& inputStream){
     //cout << questions::fullQuestion << "\n"; for debuging
 }
 
+// gets the questions
 void questions::setQuestion(){
     questions::question.clear();
-    int i = 0;
+    size_t i = 0;
     
     while(i < questions::fullQuestion.size() && questions::fullQuestion[i] != '?'){
         questions::question += questions::fullQuestion[i];
@@ -32,14 +34,15 @@ void questions::setQuestion(){
     //cout << questions::question << "\n"; for debuging
 }
 
+// gets the options and the answer
 void questions::setOptionsAndAnswer(){
     
     questions::options.clear();
 
     string nextOption;
-    int i = questions::question.size()+1; // +1 for å ikke ta med ?
+    size_t i = questions::question.size()+1; // +1 to not include ?
     
-    while(i < questions::fullQuestion.size()-1) { // -1 for å ikke ta med svaret
+    while(i < questions::fullQuestion.size()-1) { // -1 to exclude the answer
         if(questions::fullQuestion[i] != ';'){
             nextOption += questions::fullQuestion[i];
         }
@@ -49,9 +52,10 @@ void questions::setOptionsAndAnswer(){
         }
         i +=1;
     }
-    questions::answer = fullQuestion[i]-48;
+    questions::answer = fullQuestion[i]-48; // -48 to get char to int
 }
 
+// put it all togheter, less in main
 void questions::setQuestionOptionsAnswer(ifstream& inputStream){
     questions::setFullQuestion(inputStream);
     questions::setQuestion();
@@ -68,20 +72,22 @@ vector<string> questions::getOptions(){
     return questions::options;
 }
 
-int questions::getAnswer(){
+size_t questions::getAnswer(){
     return questions::answer;
 }
 
+// overload for printing the question, options and answer, for debuging
 ostream& operator << (ostream& os, const questions& p){
     
     os << "question: " << p.question << "options: " << "\n";
     
-    for(int i = 0; i < p.options.size(); i++){
+    for(size_t i = 0; i < p.options.size(); i++){
         cout << p.options[i] << " ";
     }
     return os;
 }
 
+// checks if the file is finished, by checking next char with peek
 bool questions::isFinished(ifstream& inputStream){
     return inputStream.peek() == EOF;
 }
